+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * `0.1.2`
   * Improved documentation
   * Run `npm audit fix`, upgraded dependencies
+  * Added function `sendDirectMessageToUser`
 
 * `0.1.1`
   * Improved documentation and added dev environment instructions